seperate app.jsx into components
