+++ frontend/src/App.css
@@ -3,46 +3,3 @@
 	padding: 0;
 	/* background-color: grey; */
 }
-
-#canvas {
-	padding: 0;
-	margin: 1em auto 0;
-	background-color: white;
-	display: block;
-	border: 1px solid black;
-	/* overflow: hidden; */
-}
-
-.colorContainer {
-	display: flex;
-	justify-content: center;
-}
-
-.colorSelector {
-	overflow: hidden;
-	width: fit-content;
-	display: flex;
-	justify-content: center;
-	height: 3em;
-	margin: 1em 0;
-	outline: 1px solid black;
-	border-radius: 5em;
-}
-
-.color {
-	width: 5em;
-	outline: 1px solid black;
-	margin: 0 2px;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	font-weight: 900;
-}
-
-.timeoutMessage {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	font-size: 2em;
-	margin-top: 2em;
-}
