+++ frontend/src/Canvas.css
@@ -1,0 +1,8 @@
+#canvas {
+	padding: 0;
+	margin: 1em auto 0;
+	background-color: white;
+	display: block;
+	border: 1px solid black;
+	/* overflow: hidden; */
+}
