+++ frontend/src/ColorPicker.css
@@ -1,0 +1,25 @@
+.colorContainer {
+	display: flex;
+	justify-content: center;
+}
+
+.colorSelector {
+	overflow: hidden;
+	width: fit-content;
+	display: flex;
+	justify-content: center;
+	height: 3em;
+	margin: 1em 0;
+	outline: 1px solid black;
+	border-radius: 5em;
+}
+
+.color {
+	width: 5em;
+	outline: 1px solid black;
+	margin: 0 2px;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	font-weight: 900;
+}
