+++ frontend/src/Timer.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import "./Timer.css";
+
+function Timer() {
+	return (
+		<div className="timeoutMessage">
+			Please Note that there is a 15 second cooldown between each tile draw.
+		</div>
+	);
+}
+
+export default Timer;
